+++ CHANGELOG.md
@@ -9,6 +9,8 @@ use patch releases for compatibility fixes instead.
 
 ## Unreleased
 
+## [v1.19.0](https://github.com/allenai/beaker-py/releases/tag/v1.19.0) - 2023-07-17
+
 ### Added
 
 - Added support for [Pydantic V2](https://docs.pydantic.dev/2.0/migration/).